# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DevResponse API",
    "version": "1.0.0",
    "description": "Machine-facing REST surface (`/api/v1`). Authenticate with an API key (`drk_…`) or a JWT access token as `Authorization: Bearer …`. Scopes are the application permission keys; a credential can never exceed its owner's authority. Errors use RFC 7807 `application/problem+json`; an unexpected server fault is a `500` with code `internal_error`. Every response carries an `x-request-id` correlation header, except the public, cacheable `/jwks.json` and `/openapi.json` documents."
  },
  "servers": [
    {
      "url": "https://app.devresponse.com/api/v1"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "oauth2ClientCredentials": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API key (`drk_…`) or a JWT access token."
      },
      "oauth2ClientCredentials": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://app.devresponse.com/api/v1/auth/token",
            "scopes": {
              "admin.users.read": "admin.users.read",
              "admin.users.create": "admin.users.create",
              "admin.users.update": "admin.users.update",
              "admin.users.delete": "admin.users.delete",
              "admin.users.manage": "admin.users.manage",
              "admin.users.ban": "admin.users.ban",
              "admin.users.setRole": "admin.users.setRole",
              "admin.users.setPassword": "admin.users.setPassword",
              "admin.users.sessions": "admin.users.sessions",
              "admin.users.impersonate": "admin.users.impersonate",
              "admin.roles.read": "admin.roles.read",
              "admin.roles.create": "admin.roles.create",
              "admin.roles.update": "admin.roles.update",
              "admin.roles.delete": "admin.roles.delete",
              "admin.roles.assign": "admin.roles.assign",
              "admin.groups.read": "admin.groups.read",
              "admin.groups.create": "admin.groups.create",
              "admin.groups.update": "admin.groups.update",
              "admin.groups.delete": "admin.groups.delete",
              "admin.groups.assign": "admin.groups.assign",
              "admin.permissions.manage": "admin.permissions.manage",
              "admin.orgs.read": "admin.orgs.read",
              "admin.orgs.create": "admin.orgs.create",
              "admin.orgs.update": "admin.orgs.update",
              "admin.orgs.delete": "admin.orgs.delete",
              "admin.orgs.manage": "admin.orgs.manage",
              "admin.apps.read": "admin.apps.read",
              "admin.apps.manage": "admin.apps.manage",
              "admin.audit.read": "admin.audit.read",
...
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        },
        "description": "1-indexed page number."
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 25
        },
        "description": "Rows per page (clamped to 1–200)."
      },
      "Sort": {
        "name": "sort",
        "in": "query",
        "required": false,
        "explode": true,
        "style": "form",
        "schema": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "description": "Sort directives as `field.asc` / `field.desc`, applied in order."
      },
      "Q": {
        "name": "q",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Case-insensitive full-text search."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid credential",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient scope or permission",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "RateLimited": {
        "description": "Too many requests",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "description": "RFC 7807 problem object.",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ]
      },
      "SortSpec": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "direction": {
            "type": "string",
            "enum": [
              "asc",
              "desc"
            ]
          }
        },
        "required": [
          "field",
          "direction"
        ]
      },
...
      "UserListItem": {
        "type": "object",
        "description": "A user row (snake_case, as stored).",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "better_auth_user_id": {
            "type": "string"
          },
          "primary_email": {
            "type": "string",
            "format": "email"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "pending_approval",
              "blocked",
              "suspended",
              "deactivated"
            ]
          },
          "preferred_locale": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "primary_email",
          "status",
          "created_at",
          "updated_at"
        ]
      },
...
      "CreateUserRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "user"
            ]
          },
          "initialAppStatus": {
            "type": "string",
            "enum": [
              "active",
              "pending_approval"
            ]
          },
          "preferredLocale": {
            "type": "string",
            "enum": [
              "en",
              "fr",
              "es",
              "uk",
              "pt",
              "zh",
              "hi",
              "ja"
            ]
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserCreated": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "betterAuthUserId": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "pending_approval",
              "blocked",
              "suspended",
              "deactivated"
            ]
          }
        },
        "required": [
          "id",
          "betterAuthUserId",
          "email",
          "status"
        ]
      },
...
      "UserList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserListItem"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortSpec"
            }
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ]
      },
...
    }
  },
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "listUsers",
        "tags": [
          "Users"
        ],
        "summary": "List users",
        "security": [
          {
            "bearerAuth": [
              "admin.users.read"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/Sort"
          },
          {
            "$ref": "#/components/parameters/Q"
          },
          {
            "name": "filter[status]",
            "in": "query",
            "required": false,
            "explode": true,
            "style": "form",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "active",
                  "pending_approval",
                  "blocked",
                  "suspended",
                  "deactivated"
                ]
              }
            },
            "description": "Exact-match filter; repeat the parameter for multiple values."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "operationId": "createUser",
        "tags": [
          "Users"
        ],
        "summary": "Create a user",
        "security": [
          {
            "bearerAuth": [
              "admin.users.create"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreated"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
  },
...
}